requirements.txt
winreg-unicode==0.5.0
